requirements.txt
Django==4.2
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
phonenumbers==8.13.53
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
uuid==1.30
django-cors-headers==4.6.0
celery==5.4.0
redis==5.2.1
coverage==7.6.10
flake8==7.1.1
